Removes some bootstrap. WIP
