+++ package.json
@@ -18,7 +18,6 @@
     "@emotion/styled": "^11.10.0",
     "@prisma/client": "^4.3.1",
     "@types/jsonwebtoken": "^8.5.9",
-    "bootstrap": "^5.1.3",
     "cross-fetch": "^3.1.5",
     "formik": "^2.2.6",
     "framer-motion": "^7.2.1",
@@ -32,7 +31,6 @@
     "pixi-viewport": "^4.31.0",
     "pixi.js": "^6.0.4",
     "react": "^18.2.0",
-    "react-bootstrap": "2.1.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.2.0",
     "reflect-metadata": "^0.1.13",
@@ -43,7 +41,6 @@
   },
   "devDependencies": {
     "@popperjs/core": "^2.9.1",
-    "@types/bootstrap": "^5.1.2",
     "@types/jest": "^26.0.20",
     "@types/js-yaml": "^4.0.5",
     "@types/node": "^14.14.35",
