+++ src/components/reroll/images/index.ts
@@ -1,5 +1,5 @@
 export * from "components/reroll/images/forms";
-export * from "components/reroll/images/DetailsModal";
+export * from "components/reroll/images/FileDetails";
 export * from "components/reroll/images/LibraryImageList";
 export * from "components/reroll/images/ImageList";
 export * from "components/reroll/images/UploadImageModal";
